backend/requirements.txt
flask
flask-cors
openai
pinecone-client>=3.0.0
pandas
pdfplumber
python-docx
openpyxl
python-dotenv
